Reverted because it breaks LED notifications. In particular, those which are received when the screen is on such as missed calls.
The logic doesn't make sense. Instead of using mLedNotification == null to disable the notification, it is using !mPendingPulseNotification, but if the screen is on at any point mPendingPulseNotification is set to false, which disables the notification before it has even happened.
Whatever they were trying to do by writing this code, it doesn't work!

Revert "Only pulse notification light if a new notification has been received since the screen was last turned off"

This reverts commit 2117f6fa3880e55e6a90fbd8b3a12d469c9b738b.